requirements.txt
# Core LangChain and AI
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0

# AI Providers
openai>=1.0.0
groq>=0.4.0
google-generativeai>=0.3.0
anthropic>=0.25.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles>=23.0.0
httpx>=0.25.0

# Data Processing and Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0